+++ src/components/app/app.module.css
@@ -12,9 +12,3 @@
     'feed-details feed-details'
 }
 
-.modal {
-  padding-top: 120px;
-  grid-area: ingredient;
-}
-
-
