+++ src/pages/feed-main-page/feed-main-page.module.css
@@ -1,0 +1,7 @@
+.main {
+  display: grid;
+  grid-template-columns: repeat(2, 1fr);
+  gap: 0 40px;
+  grid-template-areas: "orders stats";
+  grid-area: main;
+}
