+++ src/pages/profile/profile.jsx
@@ -4,19 +4,15 @@ import { useDispatch, useSelector } from "react-redux";
 import { logoutFromAccount } from "../../services/actions/auth";
 import { memo, useCallback, useEffect } from "react";
 import { deleteCookie } from "../../utils/cookie";
-import {
-  WS_CONNECTION_START,
-  WS_CONNECTION_CLOSED,
-} from "../../services/actions/ws-actions";
-import { getCookie } from "../../utils/cookie";
-import ApiLoader from "../../components/api-loader/api-loader";
+import { getData } from "../../services/actions/auth";
 
 function Profile({ children }) {
   const dispatch = useDispatch();
   const history = useHistory();
-  const isOrderArrayRefreshed = useSelector(
-    (state) => state.wsReducer.isRefreshed
-  );
+
+  useEffect(() => {
+    dispatch(getData());
+  }, []);
 
   const logout = useCallback((e) => {
     e.preventDefault();
@@ -26,19 +22,7 @@ function Profile({ children }) {
     history.push("/login");
   }, []);
 
-  useEffect(() => {
-    dispatch({
-      type: WS_CONNECTION_START,
-      payload: {
-        add: `?token=${getCookie("token")}`,
-      },
-    });
-    return () => {
-      dispatch({ type: WS_CONNECTION_CLOSED });
-    };
-  }, [dispatch, children]);
-
-  return isOrderArrayRefreshed ? (
+  return (
     <div className={ProfileStyles.container}>
       <div className={ProfileStyles.linksContainer}>
         <NavLink
@@ -73,8 +57,6 @@ function Profile({ children }) {
 
       <>{children}</>
     </div>
-  ) : (
-    <ApiLoader />
   );
 }
